requirement.txt
astroid==2.0.4
attrdict==2.0.0
autopep8==1.3.5
certifi==2018.8.24
chardet==3.0.4
cytoolz==0.9.0.1
defusedxml==0.5.0
Django==2.1
eth-abi==1.2.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==1.3.0
eth-utils==1.2.2
-e git+https://gitlab.com/pycqa/flake8@9631dac52aa6ed8a3de9d0983c3c7b0267ae7d6d#egg=flake8
hexbytes==0.1.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
lru-dict==1.1.6
mccabe==0.6.1
mysqlclient==1.3.13
oauthlib==2.1.0
parsimonious==0.8.0
py-solc==3.1.0
pycodestyle==2.4.0
pycryptodome==3.6.6
pyflakes==2.0.0
pylint==2.1.1
pytz==2018.5
requests==2.19.1
rlp==1.0.3
semantic-version==2.6.0
six==1.11.0
requests-oauthlib==1.0.0
rlp==1.0.3
semantic-version==2.6.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
toolz==0.9.0
urllib3==1.23
web3==4.7.2
websockets==6.0
wrapt==1.10.11
